refactor(git): extract inspection commands into git_inspect.py
