+++ .rrt/tree.lock.toml
@@ -1,10 +1,10 @@
 [meta]
-generated_at = "2026-07-05T07:31:39+00:00"
+generated_at = "2026-07-05T07:42:48+00:00"
 rrt_version = "1.10.1"
 
 [snapshot]
-entry_count = 321
-tree_hash = "sha256:87bce5185b25a03862dc918990d2e3a60d95b3d5606005f5eec0ab3b5e7e16f0"
-updated_at = "2026-07-05T07:31:39+00:00"
+entry_count = 323
+tree_hash = "sha256:ed568d5f915f8c6c4f105b460d5711a9afb8d7ec357a59f309e34776aa10252d"
+updated_at = "2026-07-05T07:42:48+00:00"
 ignored_count = 32
 phantom_empty_dirs = []
